Revert "NEW: sección ContactUs y actualización de la sección AboutUs"
