--- pages/ContactUs/index.tsx
@@ -1,23 +1,0 @@
-import StartContentContact from "../../app/components/ContacUs/StartContentContact";
-import Footer from "../../app/components/Footer/Footer";
-import FormContact from "../../app/components/FormContact/FormContact";
-import StartContentHome from "../../app/components/Home/StartContentHome";
-import WelcomeSection from "../../app/components/WelcomeSection/WelcomeSection";
-import RootLayout from "../../app/layout";
-
-
-
-export default function Home() {
-  return (
-     <>
-     <RootLayout>
-      <WelcomeSection>
-        <StartContentContact />
-      </WelcomeSection>
-      <FormContact />
-      <Footer/>
-     </RootLayout>
-     </>
-  );
-}
-
